Adapted importer to new ipc convention
